split out files
Update GraphQL schema SDL descriptions
