+++ .github/workspace-dep-closures.json
@@ -904,6 +904,89 @@
       "user_quota",
       "workspace-hack"
     ],
+    "complete_graph": [
+      "agent",
+      "ai_toolset",
+      "ai_usage",
+      "attachment",
+      "authentication_service_client",
+      "bot_id",
+      "call",
+      "channel_sender",
+      "channels",
+      "complete_graph",
+      "connection",
+      "connection_gateway_client",
+      "connection_gateway_models",
+      "contacts",
+      "cowlike",
+      "crm",
+      "decode_jwt",
+      "document_sub_type",
+      "email",
+      "email_db_client",
+      "email_utils",
+      "entity_access",
+      "entity_access_db_utils",
+      "entity_access_management",
+      "filter_ast",
+      "foreign_entity",
+      "frecency",
+      "graphql_common",
+      "graphql_notification",
+      "graphql_properties",
+      "graphql_soup",
+      "invite_email",
+      "ip_extractor",
+      "item_filters",
+      "last_online_tracker",
+      "macro_auth",
+      "macro_aws_config",
+      "macro_db_client",
+      "macro_env",
+      "macro_env_var",
+      "macro_event_broker",
+      "macro_event_topics",
+      "macro_middleware",
+      "macro_service_urls",
+      "macro_user_id",
+      "macro_uuid",
+      "mention_utils",
+      "model",
+      "model-entity",
+      "model-error-response",
+      "model_file_type",
+      "model_notifications",
+      "model_user",
+      "models_ai_projection",
+      "models_bulk_upload",
+      "models_dcs",
+      "models_email",
+      "models_entity_access_management",
+      "models_grouping",
+      "models_opensearch",
+      "models_pagination",
+      "models_permissions",
+      "models_properties",
+      "models_soup",
+      "models_team",
+      "non_empty",
+      "notification",
+      "prompt",
+      "properties",
+      "rate_limit",
+      "readonly_pool",
+      "remote_env_var",
+      "s3_key",
+      "share_permission_db_utils",
+      "soup",
+      "sqs_client",
+      "sqs_worker",
+      "system_properties",
+      "unfurl",
+      "user_quota",
+      "workspace-hack"
+    ],
     "connection": [
       "agent",
       "ai_toolset",
@@ -1435,6 +1518,7 @@
       "channels",
       "chat",
       "comms_db_client",
+      "complete_graph",
       "connection",
       "connection_gateway_client",
       "connection_gateway_models",
@@ -1460,6 +1544,9 @@
       "frecency",
       "fusionauth",
       "github",
+      "graphql_common",
+      "graphql_notification",
+      "graphql_properties",
       "graphql_soup",
       "invite_email",
       "ip_extractor",
@@ -2511,6 +2598,113 @@
       "models_pagination",
       "workspace-hack"
     ],
+    "graphql_common": [
+      "cowlike",
+      "filter_ast",
+      "graphql_common",
+      "macro_user_id",
+      "model-entity",
+      "workspace-hack"
+    ],
+    "graphql_notification": [
+      "ai_toolset",
+      "cowlike",
+      "decode_jwt",
+      "filter_ast",
+      "graphql_common",
+      "graphql_notification",
+      "last_online_tracker",
+      "macro_auth",
+      "macro_env",
+      "macro_env_var",
+      "macro_service_urls",
+      "macro_user_id",
+      "macro_uuid",
+      "model-entity",
+      "model-error-response",
+      "model_user",
+      "models_pagination",
+      "notification",
+      "rate_limit",
+      "remote_env_var",
+      "workspace-hack"
+    ],
+    "graphql_properties": [
+      "agent",
+      "ai_toolset",
+      "ai_usage",
+      "attachment",
+      "authentication_service_client",
+      "bot_id",
+      "call",
+      "channel_sender",
+      "channels",
+      "connection",
+      "connection_gateway_client",
+      "connection_gateway_models",
+      "contacts",
+      "cowlike",
+      "crm",
+      "decode_jwt",
+      "document_sub_type",
+      "email",
+      "email_db_client",
+      "email_utils",
+      "entity_access",
+      "entity_access_db_utils",
+      "entity_access_management",
+      "filter_ast",
+      "frecency",
+      "graphql_common",
+      "graphql_properties",
+      "invite_email",
+      "ip_extractor",
+      "item_filters",
+      "last_online_tracker",
+      "macro_auth",
+      "macro_aws_config",
+      "macro_db_client",
+      "macro_env",
+      "macro_env_var",
+      "macro_event_broker",
+      "macro_event_topics",
+      "macro_middleware",
+      "macro_service_urls",
+      "macro_user_id",
+      "macro_uuid",
+      "mention_utils",
+      "model",
+      "model-entity",
+      "model-error-response",
+      "model_file_type",
+      "model_notifications",
+      "model_user",
+      "models_ai_projection",
+      "models_bulk_upload",
+      "models_dcs",
+      "models_email",
+      "models_entity_access_management",
+      "models_opensearch",
+      "models_pagination",
+      "models_permissions",
+      "models_properties",
+      "models_soup",
+      "models_team",
+      "non_empty",
+      "notification",
+      "prompt",
+      "properties",
+      "rate_limit",
+      "remote_env_var",
+      "s3_key",
+      "share_permission_db_utils",
+      "sqs_client",
+      "sqs_worker",
+      "system_properties",
+      "unfurl",
+      "user_quota",
+      "workspace-hack"
+    ],
     "graphql_soup": [
       "agent",
       "ai_toolset",
@@ -2538,6 +2732,9 @@
       "filter_ast",
       "foreign_entity",
       "frecency",
+      "graphql_common",
+      "graphql_notification",
+      "graphql_properties",
       "graphql_soup",
       "invite_email",
       "ip_extractor",
