+++ rust/cloud-storage/Cargo.toml
@@ -15,6 +15,7 @@ members = [
   "client/cache-idb",
   "client/cache-sqlite",
   "client/cache-wasm",
+  "complete_graph",
   "connection_gateway",
   "contacts_service",
   "convert_service",
@@ -37,6 +38,9 @@ members = [
   "embedding",
   "favorites",
   "foreign_entity",
+  "graphql_common",
+  "graphql_notification",
+  "graphql_properties",
   "graphql_soup",
   "image_optimizer",
   "image_proxy_service",
