+++ rust/cloud-storage/complete_graph/Cargo.toml
@@ -1,15 +1,14 @@
 [package]
-name = "graphql_soup"
+name = "complete_graph"
 version = "0.1.0"
 edition = "2024"
 publish = false
 
 [dependencies]
+anyhow = { workspace = true }
 async-graphql = { workspace = true, features = ["dataloader"] }
 axum = { workspace = true }
 axum-extra = { workspace = true }
-chrono = { workspace = true }
-document_sub_type = { path = "../document_sub_type" }
 email = { path = "../email", default-features = false, features = [
   "axum",
   "inbound",
@@ -18,20 +17,16 @@ email = { path = "../email", default-features = false, features = [
 entity_access = { path = "../entity_access", default-features = false, features = [
   "inbound",
 ] }
-filter_ast = { path = "../filter_ast" }
 graphql_common = { path = "../graphql_common" }
 graphql_notification = { path = "../graphql_notification" }
 graphql_properties = { path = "../graphql_properties" }
-item_filters = { path = "../item_filters" }
-macro_user_id = { path = "../macro_user_id" }
-model_file_type = { path = "../model_file_type" }
+graphql_soup = { path = "../graphql_soup" }
 model_user = { path = "../model_user", features = ["axum"] }
-model-entity = { path = "../model-entity" }
-models_pagination = { path = "../models_pagination" }
-models_properties = { path = "../models_properties" }
-models_soup = { path = "../models_soup" }
-serde = { workspace = true }
-serde_json = { workspace = true }
 soup = { path = "../soup", default-features = false, features = ["ports"] }
-uuid = { workspace = true }
 workspace-hack = { version = "0.1", path = "../workspace-hack" }
+
+[dev-dependencies]
+macro_user_id = { path = "../macro_user_id" }
+models_pagination = { path = "../models_pagination" }
+tokio = { workspace = true }
+uuid = { workspace = true }
