+++ rust/cloud-storage/complete_graph/src/schema.rs
@@ -4,27 +4,17 @@ mod test;
 use std::{marker::PhantomData, sync::Arc};
 
 use async_graphql::{Context, EmptyMutation, EmptySubscription, Object, Schema};
-use axum::extract::{FromRef, FromRequestParts};
+use axum::extract::FromRef;
 use axum_extra::extract::Cached;
 use email::{
     domain::ports::{EmailService, NoOpEmailService},
-    inbound::axum::{axum_impls::MultiEmailLinkExtractor, previews_router::EmailRouterState},
+    inbound::axum::previews_router::EmailRouterState,
 };
-use entity_access::{
-    domain::{
-        models::{EntityAccessReceipt, MemberTeamRole},
-        ports::{EntityAccessService, NoOpEntityAccessService},
-    },
-    inbound::axum_extractors::OptionalMacroUserTeamExtractor,
-};
+use entity_access::domain::ports::{EntityAccessService, NoOpEntityAccessService};
+use graphql_common::extract_part;
+use graphql_soup::{SharedSoupService, SoupInput, SoupPage, resolve_soup};
 use model_user::axum_extractor::MacroUserExtractor;
-use models_pagination::TypeEraseCursor;
-use soup::domain::{
-    models::SoupRequest,
-    ports::{NoOpSoupService, SoupService},
-};
-
-use crate::{inputs::SoupInput, objects::SoupPage, request_context::GraphqlSoupRequestParts};
+use soup::domain::ports::{NoOpSoupService, SoupService};
 
 /// GraphQL Soup schema type.
 ///
@@ -55,56 +45,6 @@ impl FromRef<SchemaOnlyState> for EmailRouterState<NoOpEmailService> {
 impl FromRef<SchemaOnlyState> for Arc<NoOpEntityAccessService> {
     fn from_ref(_state: &SchemaOnlyState) -> Self {
         Arc::new(NoOpEntityAccessService)
-    }
-}
-
-/// Object-safe-ish wrapper for sharing a concrete Soup service with GraphQL.
-pub struct SharedSoupService<S>(Arc<S>);
-
-impl<S> Clone for SharedSoupService<S> {
-    fn clone(&self) -> Self {
-        Self(Arc::clone(&self.0))
-    }
-}
-
-impl<S> SharedSoupService<S> {
-    /// Create a shared Soup service wrapper.
-    pub fn new(service: Arc<S>) -> Self {
-        Self(service)
-    }
-}
-
-impl<S> SoupService for SharedSoupService<S>
-where
-    S: SoupService,
-{
-    async fn get_user_soup<T>(
-        &self,
-        req: SoupRequest<T>,
-        team_receipt: Option<EntityAccessReceipt<MemberTeamRole>>,
-    ) -> Result<soup::domain::ports::SoupOutput<T>, soup::domain::models::SoupErr>
-    where
-        SoupRequest<T>: soup::domain::models::IntoSoupReqAst,
-        T: Clone + serde::Serialize + Send,
-    {
-        self.0.get_user_soup(req, team_receipt).await
-    }
-
-    async fn get_user_soup_grouped(
-        &self,
-        req: soup::domain::models::GroupedSortRequest<'_>,
-    ) -> Result<Vec<soup::domain::models::GroupedSoupItem>, soup::domain::models::SoupErr> {
-        self.0.get_user_soup_grouped(req).await
-    }
-
-    async fn caller_tag_sets<'a>(
-        &self,
-        user_id: macro_user_id::user_id::MacroUserIdStr<'a>,
-    ) -> Result<
-        Vec<models_properties::service::property_definition_with_options::PropertyDefinitionWithOptions>,
-        soup::domain::models::SoupErr,
-    >{
-        self.0.caller_tag_sets(user_id).await
     }
 }
 
@@ -172,22 +112,6 @@ where
     build_schema_with_service(SharedSoupService::new(service))
 }
 
-/// Run an axum extractor against the request parts stored in the GraphQL
-/// context, using the embedding router state `St` also stored there.
-async fn extract_part<T, St>(ctx: &Context<'_>) -> async_graphql::Result<T>
-where
-    T: FromRequestParts<St>,
-    T::Rejection: std::fmt::Display,
-    St: Clone + Send + Sync + 'static,
-{
-    let parts = ctx.data::<GraphqlSoupRequestParts>()?;
-    let state = ctx.data::<St>()?;
-    parts
-        .extract_with_state::<T, St>(state)
-        .await
-        .map_err(|err| async_graphql::Error::new(err.to_string()))
-}
-
 #[Object]
 impl<S, E, EAS, St> SoupQueryRoot<S, E, EAS, St>
 where
@@ -226,32 +150,6 @@ where
 
     /// Fetch a page of Soup items using the existing Soup filter AST format.
     async fn soup(&self, ctx: &Context<'_>, input: SoupInput) -> async_graphql::Result<SoupPage> {
-        let Cached(MacroUserExtractor { macro_user_id, .. }) =
-            extract_part::<Cached<MacroUserExtractor>, St>(ctx).await?;
-        let Cached(MultiEmailLinkExtractor(links, _)) =
-            extract_part::<Cached<MultiEmailLinkExtractor<E>>, St>(ctx).await?;
-        let link_ids = links.into_iter().map(|link| link.id).collect();
-        let request = input.into_request(macro_user_id, link_ids)?;
-
-        // Team membership is only resolved when the query actually asks for
-        // CRM-scoped data; everything else skips the lookup entirely. The
-        // authorization itself (membership + admin role for hidden
-        // companies) is enforced by the soup domain and CRM service from
-        // the receipt.
-        let effective_filter = request.cursor.filter();
-        let team_receipt =
-            if effective_filter.requests_crm_scope() || effective_filter.requests_crm_admin() {
-                let Cached(team) = extract_part::<
-                    Cached<OptionalMacroUserTeamExtractor<MemberTeamRole, EAS>>,
-                    St,
-                >(ctx)
-                .await?;
-                team.entity_access_receipt
-            } else {
-                None
-            };
-
-        let page = self.service.get_user_soup(request, team_receipt).await?;
-        Ok(SoupPage::from(page.type_erase()))
+        resolve_soup::<S, E, EAS, St>(&self.service, ctx, input).await
     }
 }
