+++ rust/cloud-storage/graphql_common/src/entity_type.rs
@@ -1,0 +1,53 @@
+use async_graphql::Enum;
+use model_entity::EntityType;
+
+/// GraphQL representation of Soup entity types.
+#[derive(Enum, Copy, Clone, Eq, PartialEq)]
+pub enum GraphqlSoupEntityType {
+    /// Document entity.
+    Document,
+    /// Chat entity.
+    Chat,
+    /// Project entity.
+    Project,
+    /// Email thread entity.
+    EmailThread,
+    /// Channel entity.
+    Channel,
+    /// Channel thread entity.
+    ChannelThread,
+    /// Call entity.
+    Call,
+    /// CRM company entity.
+    CrmCompany,
+    /// Foreign entity.
+    ForeignEntity,
+    /// User entity.
+    User,
+    /// Team entity.
+    Team,
+    /// Static file entity.
+    StaticFile,
+    /// CRM contact entity.
+    CrmContact,
+}
+
+impl From<EntityType> for GraphqlSoupEntityType {
+    fn from(entity_type: EntityType) -> Self {
+        match entity_type {
+            EntityType::Document => Self::Document,
+            EntityType::Chat => Self::Chat,
+            EntityType::Project => Self::Project,
+            EntityType::EmailThread => Self::EmailThread,
+            EntityType::Channel => Self::Channel,
+            EntityType::ChannelMessage => Self::ChannelThread,
+            EntityType::Call => Self::Call,
+            EntityType::CrmCompany => Self::CrmCompany,
+            EntityType::ForeignEntity => Self::ForeignEntity,
+            EntityType::User => Self::User,
+            EntityType::Team => Self::Team,
+            EntityType::StaticFile => Self::StaticFile,
+            EntityType::CrmContact => Self::CrmContact,
+        }
+    }
+}
