+++ rust/cloud-storage/graphql_notification/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "graphql_notification"
+version = "0.1.0"
+edition = "2024"
+publish = false
+
+[dependencies]
+async-graphql = { workspace = true, features = ["dataloader"] }
+async-trait = { workspace = true }
+graphql_common = { path = "../graphql_common" }
+macro_user_id = { path = "../macro_user_id" }
+notification = { path = "../notification" }
+rootcause = { workspace = true }
+serde_json = { workspace = true }
+tokio = { workspace = true }
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
