+++ rust/cloud-storage/graphql_soup/src/lib.rs
@@ -2,29 +2,27 @@
 //!
 //! This crate is intentionally additive: it maps GraphQL requests onto the
 //! existing `soup` domain service without changing the existing REST API.
+//! The schema itself is composed and served by the `complete_graph` crate.
 #![deny(missing_docs)]
 
 mod inputs;
-mod loaders;
 mod objects;
-mod request_context;
-mod schema;
+mod resolvers;
 
-pub use inputs::{GraphqlSimpleSortMethod, SoupInput};
-pub use loaders::{
-    EntityNotificationsKey, EntityNotificationsLoader, EntityPropertiesKey, EntityPropertiesLoader,
-    SoupNotificationEdgeReader, SoupPropertyEdgeReader, entity_notifications_loader,
+pub use graphql_common::{GraphqlSoupEntityType, GraphqlSoupRequestParts};
+pub use graphql_notification::{
+    EntityNotificationsKey, EntityNotificationsLoader, GraphqlSoupNotification,
+    SoupNotificationEdgeReader, entity_notifications_loader,
+};
+pub use graphql_properties::{
+    EntityPropertiesKey, EntityPropertiesLoader, GraphqlSoupProperty,
+    GraphqlSoupPropertyEntityReference, GraphqlSoupPropertyValue, SoupPropertyEdgeReader,
     entity_properties_loader,
 };
+pub use inputs::{GraphqlSimpleSortMethod, SoupInput};
 pub use objects::{
     GraphqlSoupCall, GraphqlSoupChannel, GraphqlSoupChannelThread, GraphqlSoupChat,
     GraphqlSoupCrmCompany, GraphqlSoupDocument, GraphqlSoupDocumentSubType, GraphqlSoupEmailThread,
-    GraphqlSoupEntity, GraphqlSoupEntityType, GraphqlSoupForeignEntity, GraphqlSoupItem,
-    GraphqlSoupNotification, GraphqlSoupProject, GraphqlSoupProperty,
-    GraphqlSoupPropertyEntityReference, GraphqlSoupPropertyValue, SoupPage,
+    GraphqlSoupEntity, GraphqlSoupForeignEntity, GraphqlSoupItem, GraphqlSoupProject, SoupPage,
 };
-pub use request_context::GraphqlSoupRequestParts;
-pub use schema::{
-    SchemaOnlySoupSchema, SchemaOnlyState, SharedSoupSchema, SharedSoupService, SoupQueryRoot,
-    SoupSchema, build_schema, build_schema_from_arc, build_schema_with_service,
-};
+pub use resolvers::{SharedSoupService, resolve_soup};
