+++ rust/cloud-storage/workspace-hack/Cargo.toml
@@ -17,6 +17,7 @@ publish = false
 [dependencies]
 ahash = { version = "0.8" }
 aho-corasick = { version = "1" }
+async-graphql = { version = "7", features = ["dataloader"] }
 aws-config = { version = "1", features = ["behavior-version-latest"] }
 aws-credential-types = { version = "1", default-features = false, features = ["hardcoded-credentials", "test-util"] }
 aws-runtime = { version = "1", default-features = false, features = ["event-stream", "http-02x", "http-1x", "sigv4a"] }
@@ -136,6 +137,7 @@ zeroize = { version = "1", features = ["derive"] }
 [build-dependencies]
 ahash = { version = "0.8" }
 aho-corasick = { version = "1" }
+async-graphql = { version = "7", features = ["dataloader"] }
 aws-config = { version = "1", features = ["behavior-version-latest"] }
 aws-credential-types = { version = "1", default-features = false, features = ["hardcoded-credentials", "test-util"] }
 aws-runtime = { version = "1", default-features = false, features = ["event-stream", "http-02x", "http-1x", "sigv4a"] }
